refactor: extract the RP2040 core1 launcher into the polymod_core1 module
base/multicore/ moves verbatim to modules/polykybd/polymod_core1/ —
core1.c becomes the auto-compiled polymod_core1.c, and the fifo/irq
register helpers are renamed polymod_core1_fifo.h / polymod_core1_irq.h
so nothing that generic (irq.h) sits on the module include path beside
pico-sdk's hardware/irq.h. The four consumers (multicore_exec.c,
poly_keymap.c, base/fw_staging.c, doom/doom_mode.c) include
polymod_core1.h by bare name; the explicit SRC line leaves both variant
rules.mk, and listing the module in keyboard.json is the enable.

RP2040-only register code with nothing to mock, so no host suite — the
same shape as polymod_crc32 / polymod_rle. The CORE1_STACK_HWM probe
ships unchanged (readme pointer updated).
