+++ modules/polykybd/polymod_core1/polymod_core1.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "fifo.h"
+#include "polymod_core1_fifo.h"
 
 #include <stddef.h>
